LayoutTests/platform/mac/accessibility/popup-button-title.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../../../fast/js/resources/js-test-style.css">
<script>
var successfullyParsed = false;
</script>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<p><a id="link" role="button" aria-owns="popup" aria-haspopup="true" tabindex="0" href="#">Click me</a></p>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that a popup button will return the text under the element for its title.");

    if (window.accessibilityController) {
        document.getElementById("link").focus();
        var element = accessibilityController.focusedElement;

        // The element should not be ignored (it should be a group)
        shouldBe("element.title", "'AXTitle: Click me'");
    }

    successfullyParsed = true;
</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
